pyproject.toml
[tool.poetry]
name = "drone-connect"
version = "0.1.0"
description = ""
authors = ["Demetris Antoniades <deanton@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pymavlink = "^2.4.41"
paho-mqtt = "^2.1.0"
eclipse-zenoh = "^1.0.3"
pymavlog = "^0.4.2"
python-socketio = "^5.12.1"
requests = "^2.32.3"
websocket-client = "^1.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
